Refactor [#13] TimerViewFacade 제거, api 별 Facade로  변경 pr에 해당 시퀀스 다이어그램 표시
