+++ morib/src/main/java/org/morib/server/api/timerView/controller/TimerViewController.java
@@ -3,7 +3,7 @@ package org.morib.server.api.timerView.controller;
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
 import org.morib.server.api.timerView.dto.StopTimerRequestDto;
-import org.morib.server.api.timerView.service.TimerViewFacade;
+import org.morib.server.api.timerView.service.stop.timer.StopTimerFacade;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,19 +19,20 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class TimerViewController {
 
-    private final TimerViewFacade timerViewFacade;
+    private final StopTimerFacade stopTimerFacade;
+    //private final TodoFacade todoFacade; 이후 만들 예정
 
     @PostMapping("/timer/stop/{taskId}")
     public ResponseEntity<String> stopTimerAndFetchAccumulatedTime( // @AuthenticationPrincipal Long userId,
          @PathVariable Long taskId, @RequestBody StopTimerRequestDto dto){
-        timerViewFacade.afterStopTimer(taskId, dto);
+        stopTimerFacade.afterStopTimer(taskId, dto);
         return ResponseEntity.status(200).body("요청이 성공했습니다!");
     }
 
 
     @GetMapping("/timer/todo-card")
     public ResponseEntity getTodoCards(@RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate targetDate) {
-        timerViewFacade.getTodoCard(targetDate);
+       // todoFacade.getTodoCard(targetDate);
         return ResponseEntity.ok("void");
     }
