--- morib/src/main/java/org/morib/server/api/timerView/service/stop/timer/StopTimerService.java
@@ -1,5 +1,0 @@
-package org.morib.server.api.timerView.service.stop.timer;
-
-public interface StopTimerService {
-    void stop();
-}
