--- morib/src/main/java/org/morib/server/api/timerView/service/stop/timer/StopTimerServiceImpl.java
@@ -1,20 +1,0 @@
-package org.morib.server.api.timerView.service.stop.timer;
-
-import lombok.RequiredArgsConstructor;
-import org.morib.server.domain.timer.TimerOperator;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class StopTimerServiceImpl implements StopTimerService {
-
-    private final TimerOperator timerOperator;
-
-    @Override
-    public void stop() {
-        //timer가 멈췄을때 계산해주는 역할을 해야함!
-        timerOperator.aggregate();
-    }
-
-
-}
